Welcome to the first fully "functional" version of Log.Thoughts!
Added six new sections to the app, so that you can track your moods, automatic thoughts, evidence that supports the thoughts, evidence that does not support the thoughts, alternative/ balanced thoughts, and moods after filling in the chart.

To save the output of the chart, simply screenshot the last page of the app. We're working on a way to save this data in the app.