Summary page is now styled!
Styled the summary page to include titles for each of the fields, and spaced them out so they're more visible. Summary page is also now scrollable.

To save the output of the chart, simply screenshot the last page of the app. We're working on a way to save this data in the app.